// vitalia/src/app/core/interfaces/iappointment.ts
export interface Iappointment
{
  id: string;
  image: string;
  name: string;
  shortDescription: string;
  yearsOfExperience: number;
  consultationDuration: number;
  shortBio: string;
  specialties: string;
  servicesProvided: string;
  amountofFeeSession: number;
  // **************
  Image: string;
  Name: string;
  ShortDescription: string;
  YearsOfExperience: number;
  ConsultationDuration: number;
  AmoutOfFeeSession: number;
  Specialties: string;
  ServicesProvided: string;
  ShortBio: string;
}